+++ tests/integration_tests/rel_xfrm.rs
@@ -91,3 +91,17 @@ fn test_reuse_group_if() {
     let output = transform_str_default(input).unwrap();
     assert_contains!(output, expected);
 }
+
+#[test]
+fn test_rel_line_offset_with_transform() {
+    // implicit transform with relspec
+    let input = r##"
+<svg>
+  <point wh="0"/>
+  <path xy="^|h 10" id="p0" d="M0 0h10"/>
+  <circle cxy="#p0:50%" r="1"/>
+</svg>
+"##;
+    let output = transform_str_default(input).unwrap();
+    assert_contains!(output, r#"<circle cx="15" cy="0" r="1"/>"#);
+}